+++ CHANGELOG.md
@@ -14,15 +14,37 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 
 
-## Unreleased
+## [0.6.0] 2022-07-21
 
 ### Added
 
+* Added dedicated Scene objects form.
+* Added dedicated object property form.
+* Added custom tree form.
+* Added option to run multi-threading function.
+* Added sidedock options.
+* Added option to change App title.
+* Added matplotlib integration.
+
+### Changed
+
+### Removed
+
+
+## [0.5.0] 2022-03-22
+
+### Added
+
+* Added (optional) Ryven flow integration.
+* Added Vector object.
 * Separate shaders for background, model, and overlay objects.
 * Collection object for large numbers of individual objects.
 
 ### Changed
 
 * Changed blending of opacity to be based on camera distance.
+* Updated workflow to latest.
+* Reconstructed Camera class.
+* Switch build workflow to run with conda.
 
 ### Removed